tutorial/requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
googleapis-common-protos==1.70.0
grpclib==0.4.7
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
lxml==5.4.0
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
more-itertools==10.7.0
multidict==6.4.3
pillow==11.2.1
premailer==3.10.0
protobuf==5.29.2
pymongo==4.12.0
requests==2.32.3
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
typing_extensions==4.13.2
urllib3==2.4.0
viam-sdk==0.45.2
yagmail==0.15.293
